+++ DependencyInjection/LinkinSwaggerResolverExtension.php
@@ -13,11 +13,15 @@ declare(strict_types=1);
 
 namespace Linkin\Bundle\SwaggerResolverBundle\DependencyInjection;
 
+use Linkin\Bundle\SwaggerResolverBundle\DependencyInjection\Compiler\SwaggerValidatorCompilerPass;
 use Linkin\Bundle\SwaggerResolverBundle\Loader\JsonConfigurationLoader;
 use Linkin\Bundle\SwaggerResolverBundle\Loader\NelmioApiDocConfigurationLoader;
 use Linkin\Bundle\SwaggerResolverBundle\Loader\SwaggerConfigurationLoaderInterface;
 use Linkin\Bundle\SwaggerResolverBundle\Loader\SwaggerPhpConfigurationLoader;
 use Linkin\Bundle\SwaggerResolverBundle\Loader\YamlConfigurationLoader;
+use Linkin\Bundle\SwaggerResolverBundle\Merger\MergeStrategyInterface;
+use Linkin\Bundle\SwaggerResolverBundle\Validator\SwaggerValidatorInterface;
+use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\Definition\Exception\InvalidTypeException;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -42,7 +46,13 @@ class LinkinSwaggerResolverExtension extends Extension
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yaml');
 
+        $this->registerPathMergerStrategy($container, $config);
         $this->registerConfigurationLoader($container, $config);
+
+        $container
+            ->registerForAutoconfiguration(SwaggerValidatorInterface::class)
+            ->addTag(SwaggerValidatorCompilerPass::TAG)
+        ;
     }
 
     /**
@@ -118,5 +128,27 @@ class LinkinSwaggerResolverExtension extends Extension
         }
 
         throw new InvalidTypeException('Received unsupported file');
+    }
+
+    /**
+     * @param ContainerBuilder $container
+     * @param array $config
+     */
+    private function registerPathMergerStrategy(ContainerBuilder $container, array $config): void
+    {
+        $class = $container->findDefinition($config['path_merge_strategy'])->getClass();
+
+        if (!$class) {
+            $class = $config['path_merge_strategy'];
+        }
+
+        if (!is_subclass_of($class, MergeStrategyInterface::class)) {
+            throw new InvalidConfigurationException(sprintf(
+                'Parameter "path_merge_strategy" should contain class which implements "%s"',
+                MergeStrategyInterface::class
+            ));
+        }
+
+        $container->setAlias(MergeStrategyInterface::class, $config['path_merge_strategy']);
     }
 }